remove useless queries in common file
